requirements.txt
easyocr 
spacy 
opencv-python 
pillow 
matplotlib
notebook
pytesseract
opencv-python
scipy
numpy
pandas
torch
ipywidgets
ultralytics


